requirements.txt
streamlit
matplotlib
numpy
scikit-learn
google-generativeai
python-dotenv
